+++ CHANGELOG.md
@@ -9,6 +9,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
+* Added optional support to capture network headers.
 * Session Replay now captures text content in wireframes for UIKit applications.
 
 ## [2.2.1] - 2026-03-19